Add throwable and consumable item cookie

+++ include/camera.h
@@ -4,15 +4,15 @@
 #define _CAMERA_H
 
 struct {
-  float x;
-  float y;
-  float x_target;
-  float y_target;
+  i32 x;
+  i32 y;
+  i32 x_target;
+  i32 y_target;
   struct Entity* target;
   u8 has_target;
 } camera;
 
-void camera_init(float x, float y);
+void camera_init(i32 x, i32 y);
 
 void camera_update();
